requirements.txt
aiohttp==3.8.4
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
chardet==5.1.0
idna==3.4
multidict==6.0.4
Pillow==9.4.0
pyTelegramBotAPI==4.10.0
requests==2.28.2
telebot==0.0.5
typing-extensions==4.5.0
urllib3==1.26.14
yarl==1.8.2
python-decouple
